src/third_party/WebKit/LayoutTests/editing/shadow/selection-of-orphan-shadowroot-expected.txt

CONSOLE WARNING: line 16: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Nodes of the selection for an orphan shadow root should return null.

PASS selection.anchorNode is null
PASS selection.focusNode is null
PASS selection.baseNode is null
PASS selection.extentNode is null
PASS successfullyParsed is true

TEST COMPLETE
